ref: Make outcome an int
Store outcomes as integers so they will be faster to filter on in Snuba

Reason could also be an enumerated type, but because reasons can come
from so many different places (web/api.py, filters, and rate limits all
have their own reason strings), it would be somewhat of a nightmare to
try and maintain an enum that was the union of all reasons from all
sources.
